% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{aggregate_by_path}
\alias{aggregate_by_path}
\title{Agrège des parcours individuels}
\usage{
aggregate_by_path(data, cols = NULL, path = NULL, sep = "-")
}
\arguments{
\item{data}{Un dataframe avec des parcours par individu.}

\item{cols}{Un vecteur de chaînes de caractères (par défaut = NULL). Les
colonnes contenant les éléments des parcours individuels. Le nom de ces
colonnes doit être uniformisé sous la forme <prefix>_<niveau_de_profondeur>
(par exemple "depth_1", "depth_2", etc.). Utiliser la valeur NULL si les
parcours sont dans une colonne unique.}

\item{path}{Une chaîne de caractères (par défaut = NULL). Le nom de la
colonne si les parcours individuels sont contenus dans une seule colonne
qui doit être séparée. Utiliser la valeur NULL si les éléments des parcours
sont dans des colonnes distinctes.}

\item{sep}{Une chaîne de caractères (par défaut = "-"). Le séparateur
utilisé dans la colonne path.}
}
\value{
Un dataframe avec un comptage des parcours.
}
\description{
Agrège des parcours individuels.
}
