% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{choose_colors}
\alias{choose_colors}
\title{Associe des couleurs à des interactions}
\usage{
choose_colors(nodes, colors)
}
\arguments{
\item{nodes}{Un vecteur de chaînes de caractères. Les interactions.}

\item{colors}{Un vecteur de chaînes de caractères. Les couleurs.}
}
\value{
Un vecteur associant interactions et couleurs.
}
\description{
Associe des couleurs à des interactions.
}
\examples{
choose_colors(nodes = c("A", "B"), colors = c("#50BE87", "#4BB4E6"))

}
