% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{count_paths}
\alias{count_paths}
\title{Agrège les couples "source" - "cible" pour un niveau de profondeur donné}
\usage{
count_paths(data, depth, count = "count", prefix = "depth")
}
\arguments{
\item{data}{Un dataframe où chaque élément d'un parcours est stocké dans
une colonne distincte.}

\item{depth}{Un entier. Le niveau de profondeur des parcours à considérer.}

\item{count}{Une chaîne de caractères, le nom de la colonne contenant le
comptage des parcours.}

\item{prefix}{Une chaîne de caractères (par défaut = "depth"). Le préfixe
utilisé pour le nom des colonnes contenant les éléments des parcours.}
}
\value{
Un dataframe avec l'interaction source, l'interaction cible ainsi que
  le comptage et le niveau de profondeur associés.
}
\description{
Agrège les couples "source" - "cible" pour un niveau de profondeur donné.
}
