% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{create_links}
\alias{create_links}
\title{Crée les "links"}
\usage{
create_links(data, nodes, source = "source", target = "target",
  depth = "depth", node = "node", index = "index")
}
\arguments{
\item{data}{Un dataframe sous la forme interaction "source", interaction
"cible", comptage et niveau de profondeur associés.}

\item{nodes}{Un dataframe contenant les "nodes".}

\item{source}{Une chaîne de caractères (par défaut = "source"). Le nom de la
colonne contenant l'interaction source dans le dataframe 'data'.}

\item{target}{Une chaîne de caractères (par défaut = "target"). Le nom de la
colonne contenant l'interaction cible dans le dataframe 'data'.}

\item{depth}{Une chaîne de caractères (par défaut = "depth"). Le nom de la
colonne contenant la profondeur de l'interaction dans le dataframe 'data'.}

\item{node}{Une chaîne de caractères (par défaut = "node"). Le nom de la
colonne contenant les nodes dans le dataframe 'node'.}

\item{index}{Une chaîne de caractères (par défaut = "index"). Le nom de la
colonne contenant les index des nodes dans le dataframe 'node'.}
}
\value{
Un dataframe contenant les "links".
}
\description{
Crée les "links".
}
