% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{create_nodes}
\alias{create_nodes}
\title{Crée les "nodes"}
\usage{
create_nodes(data, source = "source", target = "target",
  depth = "depth")
}
\arguments{
\item{data}{Un dataframe sous la forme interaction "source", interaction
"cible", comptage et niveau de profondeur associés.}

\item{source}{Une chaîne de caractères (par défaut = "source"). Le nom de la
colonne contenant l'interaction source.}

\item{target}{Une chaîne de caractères (par défaut = "target"). Le nom de la
colonne contenant l'interaction cible.}

\item{depth}{Une chaîne de caractères (par défaut = "depth"). Le nom de la
colonne contenant la profondeur de l'interaction.}
}
\value{
Un dataframe contenant les "nodes".
}
\description{
Crée les "nodes".
}
