% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_1}
\alias{data_1}
\title{Exemple de parcours agrégés.}
\format{Un dataframe avec 10 lignes et 2 variables :
\describe{
  \item{parcours}{les parcours}
  \item{comptage}{le comptage associé}
}}
\usage{
data_1
}
\description{
Un dataset contenant des parcours agrégés dans une unique colonne.
}
\keyword{datasets}
