% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_2}
\alias{data_2}
\title{Exemple de parcours agrégés.}
\format{Un dataframe avec 10 lignes et 6 variables :
\describe{
  \item{contact_1}{le premier élément du parcours}
  \item{contact_2}{le second élément du parcours}
  \item{contact_3}{le troisième élément du parcours}
  \item{contact_4}{le quatrième élément du parcours}
  \item{contact_5}{le cinquième et dernier élément du parcours}
  \item{comptage}{le comptage associé}
}}
\usage{
data_2
}
\description{
Un dataset contenant des parcours agrégés dans des colonnes distinctes.
}
\keyword{datasets}
