% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_3}
\alias{data_3}
\title{Exemple de parcours individuels.}
\format{Un dataframe avec 42 lignes et 2 variables :
\describe{
  \item{id}{l'identifiant de l'individu}
  \item{parcours}{les parcours}
}}
\usage{
data_3
}
\description{
Un dataset contenant des parcours individuels dans une unique colonne.
}
\keyword{datasets}
