% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_4}
\alias{data_4}
\title{Exemple de parcours individuels.}
\format{Un dataframe avec 42 lignes et 2 variables :
\describe{
  \item{id}{l'identifiant de l'individu}
  \item{contact_1}{le premier élément du parcours}
  \item{contact_2}{le second élément du parcours}
  \item{contact_3}{le troisième élément du parcours}
  \item{contact_4}{le quatrième élément du parcours}
  \item{contact_5}{le cinquième et dernier élément du parcours}
}}
\usage{
data_4
}
\description{
Un dataset contenant des parcours individuels dans des colonnes distinctes.
}
\keyword{datasets}
