% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{from_aggregated_paths}
\alias{from_aggregated_paths}
\title{Crée un dataframe sous la forme interaction "source", interaction "cible",
comptage et niveau de profondeur associés}
\usage{
from_aggregated_paths(data, cols = NULL, path = NULL,
  count = "count", sep = "-", depth = NULL, fill = " ")
}
\arguments{
\item{data}{Un dataframe avec les parcours et comptages associés. Les
parcours peuvent être dans une colonne unique ou bien dans des colonnes
distinctes.}

\item{cols}{Un vecteur de chaînes de caractères (par défaut = NULL). Les
colonnes contenant les éléments des parcours agrégés. L'ordre est
important, il détermine le sens des éléments. Utiliser la valeur NULL si
les éléments des parcours sont dans une colonne unique.}

\item{path}{Une chaîne de caractères (par défaut = NULL). Le nom de la
colonne si les parcours agrégés sont contenus dans une seule colonne qui
doit être séparée. Utiliser la valeur NULL si les éléments des parcours
sont dans des colonnes distinctes.}

\item{count}{Une chaîne de caractères (par défaut = "count"). Le nom de la
colonne contenant les comptages.}

\item{sep}{Une chaîne de caractères (par défaut = "-"). Le séparateur utilisé
dans la colonne path.}

\item{depth}{Un entier (par défaut = NULL). Conserve l'ensemble des parcours
d'une profondeur donnée. Si NULL, les données sont laissées comme telles.}

\item{fill}{Une chaîne de caractères (par défaut = " "). Si l'argument depth
est utilisé, permet de spécifier la valeur remplacant les valeurs
manquantes.}
}
\value{
Un dataframe contenant l'interaction "source", l'interaction "cible",
  le comptage et le niveau de profondeur associés.
}
\description{
Crée un dataframe sous la forme interaction "source", interaction "cible",
comptage et niveau de profondeur associés.
}
\examples{
# parcours dans une colonne unique
from_aggregated_paths(data = data_1, path = "parcours",
                      count = "comptage", sep = " - ")

# parcours dans plusieurs colonnes
cols <- paste0("contact_", 1:5)
from_aggregated_paths(data = data_2, cols = cols, count = "comptage")

}
