% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{from_id_paths}
\alias{from_id_paths}
\title{Crée un dataframe sous la forme interaction "source", interaction "cible",
comptage et niveau de profondeur associés}
\usage{
from_id_paths(data, cols = NULL, path = NULL, sep = "-",
  depth = NULL, fill = " ")
}
\arguments{
\item{data}{Un dataframe avec des parcours par individu. Les parcours peuvent
être dans une colonne unique ou bien dans des colonnes distinctes.}

\item{cols}{Un vecteur de chaînes de caractères (par défaut = NULL). Les
colonnes contenant les éléments des parcours individuels. Le nom de ces
colonnes doit être uniformisé sous la forme <prefix>_<niveau_de_profondeur>
(par exemple "depth_1", "depth_2", etc.). Utiliser la valeur NULL si les
parcours sont dans une colonne unique.}

\item{path}{Une chaîne de caractères (par défaut = NULL). Le nom de la
colonne si les parcours individuels sont contenus dans une seule colonne
qui doit être séparée. Utiliser la valeur NULL si les éléments des parcours
sont dans des colonnes distinctes.}

\item{sep}{Une chaîne de caractères (par défaut = "-"). Le séparateur
utilisé dans la colonne path.}

\item{depth}{Un entier (par défaut = NULL). Conserve l'ensemble des parcours
d'une profondeur donnée. Si NULL, les données sont laissées comme telles.}

\item{fill}{Une chaîne de caractères (par défaut = " "). Si l'argument depth
est utilisé, permet de spécifier la valeur remplacant les valeurs
manquantes.}
}
\value{
Un dataframe contenant l'interaction "source", l'interaction "cible",
  le comptage et le niveau de profondeur associés.
}
\description{
Crée un dataframe sous la forme interaction "source", interaction "cible",
comptage et niveau de profondeur associés.
}
