% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{keep_paths}
\alias{keep_paths}
\title{Conserve l'ensemble des parcours d'une profondeur donnée}
\usage{
keep_paths(data, depth = NULL, prefix = "depth", fill = " ")
}
\arguments{
\item{data}{Un dataframe où chaque élément d'un parcours est stocké dans une
colonne distincte.}

\item{depth}{Un entier (par défaut = NULL). La profondeur des parcours à
considérer. Si NULL, les données sont laissées comme telles.}

\item{prefix}{Une chaîne de caractères (par défaut = "depth"). Le préfixe
utilisé pour le nom des colonnes contenant les éléments des parcours.}

\item{fill}{Une chaîne de caractères (par défaut = " "). La valeur utilisée
pour remplacer les valeurs manquantes.}
}
\value{
Un dataframe.
}
\description{
Conserve l'ensemble des parcours d'une profondeur donnée.
}
