% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{split_col}
\alias{split_col}
\title{Sépare une colonne de parcours en plusieurs colonnes}
\usage{
split_col(data, path, count, sep = "-", prefix = "depth")
}
\arguments{
\item{data}{Un dataframe avec les parcours et comptages associés.}

\item{path}{Une chaîne de caractères. Le nom de la colonne contenant les
parcours.}

\item{count}{Une chaîne de caractères. Le nom de la colonne contenant les
comptages.}

\item{sep}{Une chaîne de caractères (par défaut = "-"). Le séparateur
utilisé dans la colonne 'path'.}

\item{prefix}{Une chaîne de caractères (par défaut = "depth"). Le préfixe
utilisé pour le nom des colonnes contenant les éléments des parcours.}
}
\value{
Un dataframe où chaque élément d'un parcours est dans une colonne
  spécifique.
}
\description{
Sépare une colonne de parcours en plusieurs colonnes.
}
