% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd3.R
\name{aggregate_by_path}
\alias{aggregate_by_path}
\title{Agrège des parcours individuels}
\usage{
aggregate_by_path(data, cols = NULL, path = NULL, sep = "-")
}
\arguments{
\item{data}{Un dataframe avec des parcours par individu.}

\item{cols}{Un vecteur de chaînes de caractères (par défaut = NULL). Les
colonnes contenant les éléments des parcours individuels. L'ordre est
important, il détermine le sens des éléments du parcours. Utiliser la
valeur NULL si les éléments des parcours sont dans une colonne unique.}

\item{path}{Une chaîne de caractères (par défaut = NULL). Le nom de la
colonne si les parcours individuels sont contenus dans une seule colonne
qui doit être séparée. Utiliser la valeur NULL si les éléments des parcours
sont dans des colonnes distinctes.}

\item{sep}{Une chaîne de caractères (par défaut = "-"). Le séparateur utilisé
dans la colonne path.}
}
\value{
Un dataframe avec un comptage des parcours.
}
\description{
Agrège des parcours individuels.
}
\examples{
# parcours dans une colonne unique
aggregate_by_path(data = data_3, path = "parcours", sep = "-")

# parcours dans plusieurs colonnes
cols <- paste0("contact_", 1:5)
aggregate_by_path(data = data_4, cols = cols)

}
